third_party/blink/web_tests/platform/win/external/wpt/pointerevents/pointerevent_constructor.https-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] getCoalescedEvents()[0].target
  assert_equals: expected null but got Element node <div id="target0"></div>
[FAIL] getCoalescedEvents()[0].offsetX
  assert_equals: expected 300 but got 292
[FAIL] getCoalescedEvents()[0].offsetY
  assert_equals: expected 0 but got -121.140625
[FAIL] getCoalescedEvents()[1].target
  assert_equals: expected null but got Element node <div id="target0"></div>
[FAIL] getCoalescedEvents()[1].offsetX
  assert_equals: expected 310 but got 302
[FAIL] getCoalescedEvents()[1].offsetY
  assert_equals: expected 0 but got -121.140625
[FAIL] getPredictedEvents()[0].target
  assert_equals: expected null but got Element node <div id="target0"></div>
[FAIL] getPredictedEvents()[0].offsetX
  assert_equals: expected 320 but got 312
[FAIL] getPredictedEvents()[0].offsetY
  assert_equals: expected 0 but got -121.140625
[FAIL] getPredictedEvents()[1].target
  assert_equals: expected null but got Element node <div id="target0"></div>
[FAIL] getPredictedEvents()[1].offsetX
  assert_equals: expected 330 but got 322
[FAIL] getPredictedEvents()[1].offsetY
  assert_equals: expected 0 but got -121.140625
Harness: the test ran to completion.
